Update Transforming3d to use Transformer
Use Transformer to control the concrete transform. For example,
(image, label_image) pair should undergo the same transformation;
if use transform function in Transforming3d and the transform should
be randomized, there is no convenient way to guarantee the the pair
have the same tansform. However, using a Transformer, user can control
when to randomized the transform, therefore, the same transform can be
applied to the pair.